halo-backend/requirements.txt
# Halo Dispatch API requirements

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
openai==1.3.0
python-dotenv==1.0.0
websockets==12.0
sqlalchemy==2.0.23
mysql-connector-python
